# Use Seek to speed up read/write in sysfs