--- a/sysfs/digital_pin.go
+++ b/sysfs/digital_pin.go
@@ -1,6 +1,7 @@
 package sysfs
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"strconv"
@@ -37,6 +38,9 @@ type DigitalPin interface {
 type digitalPin struct {
 	pin   string
 	label string
+
+	value     File
+	direction File
 }
 
 // NewDigitalPin returns a DigitalPin given the pin number and an optional sysfs pin label.
@@ -53,18 +57,20 @@ func NewDigitalPin(pin int, v ...string) DigitalPin {
 	return d
 }
 
+var notExportedError = errors.New("pin has not been exported")
+
 func (d *digitalPin) Direction(dir string) error {
-	_, err := writeFile(fmt.Sprintf("%v/%v/direction", GPIOPATH, d.label), []byte(dir))
+	_, err := writeFile(d.direction, []byte(dir))
 	return err
 }
 
 func (d *digitalPin) Write(b int) error {
-	_, err := writeFile(fmt.Sprintf("%v/%v/value", GPIOPATH, d.label), []byte(strconv.Itoa(b)))
+	_, err := writeFile(d.value, []byte(strconv.Itoa(b)))
 	return err
 }
 
 func (d *digitalPin) Read() (n int, err error) {
-	buf, err := readFile(fmt.Sprintf("%v/%v/value", GPIOPATH, d.label))
+	buf, err := readFile(d.value)
 	if err != nil {
 		return 0, err
 	}
@@ -72,43 +78,103 @@ func (d *digitalPin) Read() (n int, err error) {
 }
 
 func (d *digitalPin) Export() error {
-	if _, err := writeFile(GPIOPATH+"/export", []byte(d.pin)); err != nil {
+	export, err := fs.OpenFile(GPIOPATH+"/export", os.O_WRONLY, 0644)
+	if err != nil {
+		return err
+	}
+	defer export.Close()
+
+	_, err = writeFile(export, []byte(d.pin))
+	if err != nil {
 		// If EBUSY then the pin has already been exported
 		if err.(*os.PathError).Err != syscall.EBUSY {
 			return err
 		}
 	}
-	return nil
+
+	if d.direction != nil {
+		d.direction.Close()
+	}
+
+	d.direction, err = fs.OpenFile(fmt.Sprintf("%v/%v/direction", GPIOPATH, d.label), os.O_RDWR, 0644)
+
+	if d.value != nil {
+		d.value.Close()
+	}
+	if err == nil {
+		d.value, err = fs.OpenFile(fmt.Sprintf("%v/%v/value", GPIOPATH, d.label), os.O_RDWR, 0644)
+	}
+
+	if err != nil {
+		// Should we unexport here?
+		// If we don't unexport we should make sure to close d.direction and d.value here
+		d.Unexport()
+	}
+
+	return err
 }
 
 func (d *digitalPin) Unexport() error {
-	if _, err := writeFile(GPIOPATH+"/unexport", []byte(d.pin)); err != nil {
+	unexport, err := fs.OpenFile(GPIOPATH+"/unexport", os.O_WRONLY, 0644)
+	if err != nil {
+		return err
+	}
+	defer unexport.Close()
+
+	if d.direction != nil {
+		d.direction.Close()
+		d.direction = nil
+	}
+	if d.value != nil {
+		d.value.Close()
+		d.value = nil
+	}
+
+	_, err = writeFile(unexport, []byte(d.pin))
+	if err != nil {
 		// If EINVAL then the pin is reserved in the system and can't be unexported
 		if err.(*os.PathError).Err != syscall.EINVAL {
 			return err
 		}
 	}
+
 	return nil
 }
 
-var writeFile = func(path string, data []byte) (i int, err error) {
-	file, err := OpenFile(path, os.O_WRONLY, 0644)
-	defer file.Close()
-	if err != nil {
-		return
+// Linux sysfs / GPIO specific sysfs docs.
+//  https://www.kernel.org/doc/Documentation/filesystems/sysfs.txt
+//  https://www.kernel.org/doc/Documentation/gpio/sysfs.txt
+
+var writeFile = func(f File, data []byte) (i int, err error) {
+	if f == nil {
+		return 0, notExportedError
 	}
 
-	return file.Write(data)
+	// sysfs docs say:
+	// > When writing sysfs files, userspace processes should first read the
+	// > entire file, modify the values it wishes to change, then write the
+	// > entire buffer back.
+	// however, this seems outdated/inaccurate (docs are from back in the Kernel BitKeeper days).
+
+	i, err = f.Write(data)
+	return i, err
 }
 
-var readFile = func(path string) ([]byte, error) {
-	file, err := OpenFile(path, os.O_RDONLY, 0644)
-	defer file.Close()
-	if err != nil {
-		return make([]byte, 0), err
+var readFile = func(f File) ([]byte, error) {
+	if f == nil {
+		return nil, notExportedError
 	}
 
+	// sysfs docs say:
+	// > If userspace seeks back to zero or does a pread(2) with an offset of '0' the [..] method will
+	// > be called again, rearmed, to fill the buffer.
+
+	// TODO: Examine if seek is needed if full buffer is read from sysfs file.
+
 	buf := make([]byte, 2)
-	_, err = file.Read(buf)
+	_, err := f.Seek(0, os.SEEK_SET)
+	if err == nil {
+		_, err = f.Read(buf)
+	}
 	return buf, err
 }