--- a/sysfs/digital_pin_test.go
+++ b/sysfs/digital_pin_test.go
@@ -24,53 +24,60 @@ func TestDigitalPin(t *testing.T) {
 	gobot.Assert(t, pin.label, "custom")
 
 	pin = NewDigitalPin(10).(*digitalPin)
+	gobot.Assert(t, pin.pin, "10")
 	gobot.Assert(t, pin.label, "gpio10")
+	gobot.Assert(t, pin.value, nil)
 
-	pin.Unexport()
+	err := pin.Unexport()
+	gobot.Assert(t, err, nil)
 	gobot.Assert(t, fs.Files["/sys/class/gpio/unexport"].Contents, "10")
 
-	pin.Export()
-	gobot.Assert(t, fs.Files["/sys/class/gpio/unexport"].Contents, "10")
+	err = pin.Export()
+	gobot.Assert(t, err, nil)
+	gobot.Assert(t, fs.Files["/sys/class/gpio/export"].Contents, "10")
+	gobot.Refute(t, pin.value, nil)
 
-	pin.Write(1)
+	err = pin.Write(1)
+	gobot.Assert(t, err, nil)
 	gobot.Assert(t, fs.Files["/sys/class/gpio/gpio10/value"].Contents, "1")
 
-	pin.Direction(IN)
+	err = pin.Direction(IN)
+	gobot.Assert(t, err, nil)
 	gobot.Assert(t, fs.Files["/sys/class/gpio/gpio10/direction"].Contents, "in")
 
 	data, _ := pin.Read()
 	gobot.Assert(t, 1, data)
 
 	pin2 := NewDigitalPin(30, "custom")
-	err := pin2.Write(1)
+	err = pin2.Write(1)
 	gobot.Refute(t, err, nil)
 
 	data, err = pin2.Read()
 	gobot.Refute(t, err, nil)
 	gobot.Assert(t, data, 0)
 
-	writeFile = func(string, []byte) (int, error) {
+	writeFile = func(File, []byte) (int, error) {
 		return 0, &os.PathError{Err: syscall.EINVAL}
 	}
 
 	err = pin.Unexport()
 	gobot.Assert(t, err, nil)
 
-	writeFile = func(string, []byte) (int, error) {
+	writeFile = func(File, []byte) (int, error) {
 		return 0, &os.PathError{Err: errors.New("write error")}
 	}
 
 	err = pin.Unexport()
 	gobot.Assert(t, err.(*os.PathError).Err, errors.New("write error"))
 
-	writeFile = func(string, []byte) (int, error) {
+	writeFile = func(File, []byte) (int, error) {
 		return 0, &os.PathError{Err: syscall.EBUSY}
 	}
 
 	err = pin.Export()
 	gobot.Assert(t, err, nil)
 
-	writeFile = func(string, []byte) (int, error) {
+	writeFile = func(File, []byte) (int, error) {
 		return 0, &os.PathError{Err: errors.New("write error")}
 	}