--- a/sysfs/fs.go
+++ b/sysfs/fs.go
@@ -11,6 +11,7 @@ type File interface {
 	Sync() (err error)
 	Read(b []byte) (n int, err error)
 	ReadAt(b []byte, off int64) (n int, err error)
+	Seek(offset int64, whence int) (ret int64, err error)
 	Fd() uintptr
 	Close() error
 }