@stengaard 2016-02-18
@trevrosen @deadprogram here's the rewritten version. Feedback is of course more than welcome!