@trevrosen 2016-02-18
@stengaard can you repro Travis' test failure for the Edison adapter locally?