@stengaard 2016-02-19
Yeah, I see the same problem popped up in the original PR.