@trevrosen 2016-02-19
Yeah that would be a regression... :crying_cat_face:  If you can put together a fix, pretty sure @deadprogram would be willing to merge tho!