@stengaard 2016-02-19
@trevrosen - should be fixed by f839367