@deadprogram 2016-02-19
This is really cool, @stengaard I'm plugging in a few devices to give it some "real-world" testing right now...