@deadprogram 2016-02-19
OK, so far I am extremely impressed by the performance boot. I do not have the formal benchmarks that you did @stengaard but anecdotally it is notable.

I am merging this with great pleasure, thanks @trevrosen for you help in landing it, as well.